Testing

+++ src/main/java/revature/d33gz/services/ClientServiceImplement.java
@@ -31,12 +31,26 @@ public class ClientServiceImplement implements ClientService {
 
 	//Update
 	public Client updateClient(Client client, int id) {
-		this.cdao.updateClient(client, id);
-		return client;
+		Client returnClient;
+		Client checkingClient = this.cdao.getOneClient(id);
+		if (checkingClient.getId() == 0) {
+			returnClient = checkingClient;
+		} else {
+			returnClient = this.cdao.updateClient(client, id);;
+		}
+		return returnClient;
 	}
 	
 	//Delete
-	public void deleteClient(int id) {
-		this.cdao.deleteClient(id);
+	public boolean deleteClient(int id) {
+		boolean yayOrNay;
+		Client checkingClient = this.cdao.getOneClient(id);
+		System.out.println(checkingClient);
+		if (checkingClient.getId() == 0) {
+			yayOrNay = false;
+		} else {
+			yayOrNay = this.cdao.deleteClient(id);
+		}
+		return yayOrNay;
 	}
 }
